requirements.txt
requests
googlesearch-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
